Refactor package to follow the module name
