+++ molgenis-amazon-bucket/src/test/java/org/molgenis/amazon/bucket/config/AmazonBucketTestConfig.java
@@ -1,10 +1,10 @@
-package org.molgenis.file.ingest.bucket.config;
+package org.molgenis.amazon.bucket.config;
 
+import org.molgenis.amazon.bucket.meta.AmazonBucketJobExecutionFactory;
+import org.molgenis.amazon.bucket.meta.AmazonBucketJobExecutionMetaData;
 import org.molgenis.data.config.EntityBaseTestConfig;
 import org.molgenis.data.jobs.config.JobTestConfig;
 import org.molgenis.file.config.FileTestConfig;
-import org.molgenis.file.ingest.bucket.meta.AmazonBucketJobExecutionFactory;
-import org.molgenis.file.ingest.bucket.meta.AmazonBucketJobExecutionMetaData;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
